+++ src/foreign.c
@@ -2300,7 +2300,7 @@ int igraph_write_graph_dimacs(const igraph_t *graph, FILE *outstream,
 int igraph_i_gml_convert_to_key(const char *orig, char **key) {
   static int no=1;
   char strno[50];
-  long int i, len=strlen(orig), newlen=0, plen=0;
+  size_t i, len = strlen(orig), newlen = 0, plen = 0;
   igraph_bool_t pref=0;  
   /* do we need a prefix? */
   if (len==0 || !isalpha(orig[0])) {
