+++ src/walktrap_graph.cpp
@@ -212,12 +212,12 @@ int Graph::convert_from_igraph(const igraph_t *graph,
 }
 
 long Graph::memory() {
-  long m = 0;
-  m += long(nb_vertices)*sizeof(Vertex);
-  m += 2*long(nb_edges)*sizeof(Edge);
+  size_t m = 0;
+  m += size_t(nb_vertices)*sizeof(Vertex);
+  m += 2*size_t(nb_edges)*sizeof(Edge);
   m += sizeof(Graph);
   if(index != 0) {
-    m += long(nb_vertices)*sizeof(char*);
+    m += size_t(nb_vertices)*sizeof(char*);
     for(int i = 0; i < nb_vertices; i++)
       m += strlen(index[i]) + 1;
   }
